TST/MAINT: split up test_resample.py (GH17806)
